+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Add BattleNet Provider [#260](https://github.com/Sorcery/sorcery/pull/260)
 * Fix ruby 2.7 deprecation warnings [#241](https://github.com/Sorcery/sorcery/pull/241)
+* Add Argon2 support [#266](https://github.com/Sorcery/sorcery/pull/266)
 
 ## 0.15.0